Fixing data types

+++ OperatorsAndConditions/05.PointInRectangle/Startup.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 int main()
 {
-	int x1 = 0, y1 = 0, x2 = 0, y2 = 0, x = 0, y = 0;
+	double x1 = 0, y1 = 0, x2 = 0, y2 = 0, x = 0, y = 0;
 	cin >> x1 >> y1 >> x2 >> y2 >> x >> y;
 
 	if (x >= x1 && x <= x2 && y >= y1 && y <= y2)
